fix: move parseSasViyaDebugResponse method to utils folder
