+++ src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './splitChunks'
 export * from './parseWeboutResponse'
 export * from './fetchLogByChunks'
 export * from './isValidJson'
+export * from './parseViyaDebugResponse'
